Move getCmapMapping from incrementalfontutils.js to binaryfonteditor.js. Make a typedef for CmapMapping.
